little refactor: added double-decycling
